+++ typed-protocols/src/Network/TypedProtocol/Driver.hs
@@ -19,9 +19,11 @@ import Data.Void (Void)
 import Network.TypedProtocol.Core
 import Network.TypedProtocol.Peer
 
+import Control.DeepSeq (NFData, force)
 import Control.Concurrent.Class.MonadSTM.TQueue
 import Control.Monad.Class.MonadAsync
 import Control.Monad.Class.MonadFork
+import Control.Monad.Class.MonadThrow
 import Control.Monad.Class.MonadSTM
 
 
@@ -116,7 +118,11 @@ data SomeMessage (st :: ps) where
 --
 runPeerWithDriver
   :: forall ps (st :: ps) pr dstate m a.
-     Monad m
+     ( Monad m
+     , MonadEvaluate m
+     , NFData a
+     , NFData dstate
+     )
   => Driver ps pr dstate m
   -> Peer ps pr NonPipelined st m a
   -> m (a, dstate)
@@ -128,7 +134,7 @@ runPeerWithDriver Driver{sendMessage, recvMessage, initialDState} =
        -> Peer ps pr 'NonPipelined st' m a
        -> m (a, dstate)
     go dstate (Effect k) = k >>= go dstate
-    go dstate (Done _ x) = return (x, dstate)
+    go dstate (Done _ x) = evaluate (force (x, dstate))
 
     go dstate (Yield refl msg k) = do
       sendMessage refl msg
@@ -165,18 +171,23 @@ runPeerWithDriver Driver{sendMessage, recvMessage, initialDState} =
 --
 runPipelinedPeerWithDriver
   :: forall ps (st :: ps) pr dstate m a.
-     MonadAsync m
+     ( MonadAsync m
+     , MonadEvaluate m
+     , NFData a
+     )
   => Driver ps pr dstate m
   -> PeerPipelined ps pr st m a
   -> m (a, dstate)
 runPipelinedPeerWithDriver driver@Driver{initialDState} (PeerPipelined peer) = do
     receiveQueue <- atomically newTQueue
     collectQueue <- atomically newTQueue
-    a <- runPipelinedPeerReceiverQueue receiveQueue collectQueue driver
+    r@(a, _dstate) <- runPipelinedPeerReceiverQueue receiveQueue collectQueue driver
            `withAsyncLoop`
          runPipelinedPeerSender        receiveQueue collectQueue driver
                                        peer initialDState
-    return a
+                                  
+    _ <- evaluate (force a)
+    return r
 
   where
     withAsyncLoop :: m Void -> m x -> m x